+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [CalVer](https://calver.org/).
 
 ## [UNRELEASED] - 2026-23-3
 ### Changed
-- Throw exception for unauthorized database access in `Layer` model.
+- Throw an exception for unauthorized database access in Meta v1 API. 
+  If user is not authorized to access the database, a exception with code 401 and message "USER_NOT_AUTHORIZED" is thrown.
+  Anonymous access is still possible. 
 
 ### Fixed
 - Stop session on error handling in OAuth2 authentication flow.