+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
-## [2026.3.11] - 2026-23-3
+## [2026.3.11] - 2026-30-3
 ### Changed
 - Throw an exception for unauthorized database access in Meta v1 API. 
   If a user is not authorized to access the database, an exception with code 401 and the message "USER_NOT_AUTHORIZED" is thrown.